refactor: move chord generator to /chords route, splash screen at /
